requirements.txt
astroplan
astropy
black
emcee
iminuit
matplotlib
numpy
pandas
pytest
pyyaml
scipy

# model-specific
sncosmo
dustmaps
dust_extinction

# messengers
python-telegram-bot
slack_sdk

# brokers
fink-client
fastavro
lasair
alerce

# web-tools
jinja2
